Add a mixin to ease adding new palettes
