+++ qdarkstyle/darkpalette.py
@@ -3,14 +3,12 @@
 
 """QDarkStyle default dark palette."""
 
-# Standard library imports
-from collections import OrderedDict
-
 # Local imports
 from qdarkstyle.colorsystem import Blue, Gray
+from qdarkstyle.utils.palette import PaletteMixin
 
 
-class DarkPalette(object):
+class DarkPalette(PaletteMixin):
     """Theme variables."""
 
     # Color
@@ -46,56 +44,8 @@ class DarkPalette(object):
     BORDER_SELECTION_2 = '1px solid $COLOR_ACCENT_2'
     BORDER_SELECTION_1 = '1px solid $COLOR_ACCENT_1'
 
-
     # Example of additional widget specific variables
     W_STATUS_BAR_BACKGROUND_COLOR = COLOR_ACCENT_1
 
     # Paths
     PATH_RESOURCES = "':/qss_icons'"
-
-    @classmethod
-    def to_dict(cls, colors_only=False):
-        """Convert variables to dictionary."""
-        order = [
-            'COLOR_BACKGROUND_6',
-            'COLOR_BACKGROUND_5',
-            'COLOR_BACKGROUND_4',
-            'COLOR_BACKGROUND_2',
-            'COLOR_BACKGROUND_3',
-            'COLOR_BACKGROUND_1',
-            'COLOR_TEXT_1',
-            'COLOR_TEXT_2',
-            'COLOR_TEXT_3',
-            'COLOR_TEXT_4',
-            'COLOR_ACCENT_1',
-            'COLOR_ACCENT_2',
-            'COLOR_ACCENT_3',
-            'COLOR_ACCENT_4',
-            'OPACITY_TOOLTIP',
-            'SIZE_BORDER_RADIUS',
-            'BORDER_1',
-            'BORDER_2',
-            'BORDER_3',
-            'BORDER_SELECTION_3',
-            'BORDER_SELECTION_2',
-            'BORDER_SELECTION_1',
-            'W_STATUS_BAR_BACKGROUND_COLOR',
-            'PATH_RESOURCES',
-        ]
-        dic = OrderedDict()
-        for var in order:
-            value = getattr(cls, var)
-
-            if colors_only:
-                if not var.startswith('COLOR'):
-                    value = None
-
-            if value:
-                dic[var] = value
-
-        return dic
-
-    @classmethod
-    def color_palette(cls):
-        """Return the ordered colored palette dictionary."""
-        return cls.to_dict(colors_only=True)
